refactor(generators): update page blueprint
