+++ .vuesion/generators/component/component.spec.ts.hbs
@@ -1,9 +1,17 @@
-import { render } from '@testing-library/vue';
+import { render, RenderResult } from '@testing-library/vue';
 import {{ properCase componentName }} from './{{ properCase componentName }}.vue';
 
 describe('{{ properCase componentName }}.vue', () => {
+  let harness: RenderResult;
+
+  beforeEach(() => {
+    harness = render({{ properCase componentName }}, {
+      stubs: ['nuxt-link'],
+    });
+  });
+
   test('renders component', () => {
-    const { getByText } = render(VueText);
+    const { getByText } = harness;
 
     getByText('{{ properCase componentName }}');
   });
