+++ src/components/molecules/VuePagination/VuePagination.stories.ts
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/vue';
 import { action } from '@storybook/addon-actions';
-import { i18n } from '@/test/test-utils';
+import { i18n } from '@/test/i18n';
 import VuePagination from './VuePagination.vue';
 
 const story = storiesOf('Molecules|Pagination', module) as any;
