+++ src/components/molecules/VueTruncate/VueTruncate.stories.ts
@@ -1,5 +1,5 @@
 import { storiesOf } from '@storybook/vue';
-import { i18n } from '@/test/test-utils';
+import { i18n } from '@/test/i18n';
 import VueTruncate from './VueTruncate.vue';
 
 const story = storiesOf('Molecules|Truncate', module) as any;
