+++ src/components/organisms/FormExample/FormExample.spec.ts
@@ -1,5 +1,5 @@
 import { createLocalVue, mount } from '@vue/test-utils';
-import { i18n } from '@/test/test-utils';
+import { i18n } from '@/test/i18n';
 import FormExample from './FormExample.vue';
 
 const localVue = createLocalVue();
