+++ src/components/organisms/VueAutocomplete/VueAutocomplete.spec.ts
@@ -1,5 +1,6 @@
 import { createLocalVue, mount, shallowMount } from '@vue/test-utils';
-import { i18n, triggerDocument } from '@/test/test-utils';
+import { i18n } from '@/test/i18n';
+import { triggerDocument } from '@/test/test-utils';
 import VueLoader from '../../atoms/VueLoader/VueLoader.vue';
 import VueAutocomplete from './VueAutocomplete.vue';
 import { AutocompleteFixture } from './IAutocompleteFixture';
