+++ src/components/organisms/VueCalendar/VueCalendar.spec.ts
@@ -1,5 +1,5 @@
 import { createLocalVue, mount } from '@vue/test-utils';
-import { i18n } from '@/test/test-utils';
+import { i18n } from '@/test/i18n';
 import VueCalendar from './VueCalendar.vue';
 
 const localVue = createLocalVue();
