+++ src/components/organisms/VueDataTable/VueDataTable.spec.ts
@@ -1,5 +1,5 @@
 import { createLocalVue, mount } from '@vue/test-utils';
-import { i18n } from '@/test/test-utils';
+import { i18n } from '@/test/i18n';
 import VuePagination from '../../molecules/VuePagination/VuePagination.vue';
 import { IComputedDataRowCell, IDataTableHeader } from './IDataTable';
 import VueDataTable from './VueDataTable.vue';
