+++ src/components/organisms/VueDatePicker/VueDatePicker.stories.ts
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/vue';
 import { action } from '@storybook/addon-actions';
-import { i18n } from '@/test/test-utils';
+import { i18n } from '@/test/i18n';
 import VueDatePicker from './VueDatePicker.vue';
 
 const story = storiesOf('Organisms|DatePicker', module) as any;
