+++ src/components/organisms/VueDateRangePicker/VueDateRangePicker.stories.ts
@@ -1,6 +1,6 @@
 import { storiesOf } from '@storybook/vue';
 import { action } from '@storybook/addon-actions';
-import { i18n } from '@/test/test-utils';
+import { i18n } from '@/test/i18n';
 import VueDateRangePicker from './VueDateRangePicker.vue';
 
 const story = storiesOf('Organisms|DateRangePicker', module) as any;
