+++ src/interfaces/IItem.ts
