+++ src/test/jestsetup.ts
@@ -3,6 +3,7 @@ import Vuex from 'vuex';
 import VueCompositionApi from '@vue/composition-api';
 import VueI18n from 'vue-i18n';
 import '@/plugins/vee-validate/vee-validate.ts';
+import { i18n } from '@/test/i18n';
 
 Vue.use(Vuex);
 Vue.use(VueCompositionApi);
@@ -20,3 +21,11 @@ Vue.mixin({
     this.$style = $style;
   },
 });
+
+Vue.prototype.$nuxt = {
+  context: {
+    app: {
+      i18n,
+    },
+  },
+};
