+++ src/test/test-utils.ts
@@ -1,8 +1,6 @@
 import axios, { AxiosRequestConfig } from 'axios';
 import MockAdapter from 'axios-mock-adapter';
-import VueI18n from 'vue-i18n';
 import { NuxtAxiosInstance } from '@nuxtjs/axios';
-import { dateTimeFormats } from '@/plugins/vue-i18n/vue-i18n';
 
 export const TestComponent = (setup: (props?: any, options?: any) => void, template = '<div/>'): any => ({
   setup,
@@ -29,14 +27,6 @@ if (typeof beforeEach !== 'undefined') {
     document.removeEventListener = removeEventListener;
   });
 }
-
-export const i18n = new VueI18n({
-  dateTimeFormats,
-});
-
-i18n.locale = 'en';
-i18n.fallbackLocale = 'en';
-i18n.setLocaleMessage('en', require('../../i18n/en.json'));
 
 export interface AxiosMock extends NuxtAxiosInstance, MockAdapter {}
